Schedule jobs through the Azkaban API.
The zip uploader will now search for a json file in the same location as
the project zip files that has the same name (except for a '.json'
extension). If found, this file is expected to contain a cron
schedule for each environment of each flow within the job (though
neither all the flows nor all the environments have to have one).

If a schedule is found for a flow for the environment the lambda is
deployed to, then the azkaban endpoint to schedule a flow is invoked.

An example schedule file is given in
'resources/aws-payment-timelines.json'.